Move src/convert.ts to src/lib/
