+++ src/background.ts
@@ -7,7 +7,7 @@ import "./lib/browser_proxy_background"
 import * as messaging from "./messaging"
 import * as excmds from "./.excmds_background.generated"
 import * as commandline_background from "@src/background/commandline_background"
-import * as convert from "./convert"
+import * as convert from "@src/lib/convert"
 import * as config from "@src/lib/config"
 import * as dom from "./dom"
 import * as download_background from "./download_background"
